ISSUE-151: Reorganize Slack Infra adapter
