+++ config/routes.yaml
@@ -5,7 +5,7 @@
 
 chat_slack_tr:
     path: /chat/slack/tr
-    controller: slub.infrastructure.chat.slack.tr_action::executeAction
+    controller: slub.infrastructure.chat.slack.tr.ok_action::executeAction
     methods: [POST]
 
 chat_slack_oauth_installation:
