+++ config/services.yaml
@@ -152,14 +152,13 @@ services:
         arguments:
             - '@slub.infrastructure.chat.slack.slub_bot'
 
-    slub.infrastructure.chat.slack.tr_action: #Consider renaming register TR submission
-        class: Slub\Infrastructure\Chat\Slack\TRAction
+    slub.infrastructure.chat.slack.tr.ok_action: #Consider renaming register TR submission
+        class: Slub\Infrastructure\Chat\Slack\TR\OKAction
 
-    slub.infrastructure.chat.slack.process_tr_action:
-        class: Slub\Infrastructure\Chat\Slack\ProcessTRRequestListener
+    slub.infrastructure.chat.slack.tr.process_tr_async:
+        class: Slub\Infrastructure\Chat\Slack\TR\ProcessTRAsync
         arguments:
             - '@slub.application.put_pr_to_review.put_pr_to_review_handler'
-            - '@slub.infrastructure.query.get_channel_information'
             - '@slub.infrastructure.vcs.github.query.get_pr_info'
             - '@slub.infrastructure.chat.slack.slack_client'
             - '@router'
@@ -168,7 +167,7 @@ services:
             - { name: kernel.event_listener, event: kernel.terminate }
 
     slub.infrastructure.chat.slack.oauth_installation_action:
-        class: Slub\Infrastructure\Chat\Slack\OAuthInstallationAction
+        class: Slub\Infrastructure\Chat\Slack\AppInstallation\OAuthInstallationAction
         arguments:
             - '@slub.infrastructure.http.guzzle_client'
             - '@slub.infrastructure.chat.slack.slack_app_installation_repository'
@@ -177,20 +176,20 @@ services:
             - '%env(GITHUB_APP_HOMEPAGE_URL)%'
 
     slub.infrastructure.query.get_channel_information:
-        class: Slub\Infrastructure\Chat\Slack\GetChannelInformation
+        class: Slub\Infrastructure\Chat\Slack\Query\GetChannelInformation
         arguments:
             - '@slub.infrastructure.http.guzzle_client'
             - '@slub.infrastructure.chat.slack.slack_app_installation_repository'
 
     slub.infrastructure.query.get_bot_reactions_for_message_and_user:
-        class: Slub\Infrastructure\Chat\Slack\GetBotReactionsForMessageAndUser
+        class: Slub\Infrastructure\Chat\Slack\Query\GetBotReactionsForMessageAndUser
         arguments:
             - '@slub.infrastructure.http.guzzle_client'
             - '@slub.infrastructure.chat.slack.slack_app_installation_repository'
             - '@logger'
 
     slub.infrastructure.query.get_bot_user_id:
-        class: Slub\Infrastructure\Chat\Slack\GetBotUserId
+        class: Slub\Infrastructure\Chat\Slack\Query\GetBotUserId
         arguments:
             - '@slub.infrastructure.http.guzzle_client'
             - '@slub.infrastructure.chat.slack.slack_app_installation_repository'
