+++ src/Slub/Infrastructure/Chat/Slack/AppInstallation/SlackAppInstallation.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Slub\Infrastructure\Chat\Slack;
+namespace Slub\Infrastructure\Chat\Slack\AppInstallation;
 
 use ConvenientImmutability\Immutable;
