+++ src/Slub/Infrastructure/Chat/Slack/Common/APIHelper.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Slub\Infrastructure\Chat\Slack;
+namespace Slub\Infrastructure\Chat\Slack\Common;
 
 use Psr\Http\Message\ResponseInterface;
