+++ src/Slub/Infrastructure/Chat/Slack/Common/ImpossibleToParseRepositoryURL.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Slub\Infrastructure\Chat\Slack\Common;
+
+/**
+ * @author    Samir Boulil <samir.boulil@gmail.com>
+ */
+class ImpossibleToParseRepositoryURL extends \InvalidArgumentException
+{
+}
