--- src/Slub/Infrastructure/Chat/Slack/ImpossibleToParseGithubURL.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Slub\Infrastructure\Chat\Slack;
-
-/**
- * @author    Samir Boulil <samir.boulil@gmail.com>
- */
-class ImpossibleToParseGithubURL extends \InvalidArgumentException
-{
-}
