+++ src/Slub/Infrastructure/Chat/Slack/Query/GetBotUserIdInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Slub\Infrastructure\Chat\Slack;
+namespace Slub\Infrastructure\Chat\Slack\Query;
 
 /**
  * @author    Samir Boulil <samir.boulil@gmail.com>
