+++ tests/Integration/Infrastructure/Chat/Slack/ChannelIdentifierHelperTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Tests\Integration\Infrastructure\Chat\Slack;
 
 use PHPUnit\Framework\TestCase;
-use Slub\Infrastructure\Chat\Slack\ChannelIdentifierHelper;
+use Slub\Infrastructure\Chat\Slack\Common\ChannelIdentifierHelper;
 
 /**
  * @author    Samir Boulil <samir.boulil@gmail.com>
@@ -17,7 +17,7 @@ class ChannelIdentifierHelperTest extends TestCase
      */
     public function it_creates_a_string_out_of_a_workspace_and_channel(): void
     {
-        $this->assertEquals('akeneo@general', ChannelIdentifierHelper::from('akeneo', 'general'));
+        $this->assertEquals('akeneo@general', \Slub\Infrastructure\Chat\Slack\Common\ChannelIdentifierHelper::from('akeneo', 'general'));
     }
 
     /**
@@ -25,7 +25,7 @@ class ChannelIdentifierHelperTest extends TestCase
      */
     public function it_returns_the_channel_and_ts_out_of_a_normalized_channel_identifier(): void
     {
-        $this->assertEquals(['workspace' => 'akeneo', 'channel' => 'general'], ChannelIdentifierHelper::split('akeneo@general'));
+        $this->assertEquals(['workspace' => 'akeneo', 'channel' => 'general'], \Slub\Infrastructure\Chat\Slack\Common\ChannelIdentifierHelper::split('akeneo@general'));
     }
 
     /**
@@ -34,6 +34,6 @@ class ChannelIdentifierHelperTest extends TestCase
     public function it_throws_if_the_channel_identifier_is_malformed(): void
     {
         $this->expectException(\InvalidArgumentException::class);
-        ChannelIdentifierHelper::split('channel_identifier_without_separator');
+        \Slub\Infrastructure\Chat\Slack\Common\ChannelIdentifierHelper::split('channel_identifier_without_separator');
     }
 }
