+++ tests/Integration/Infrastructure/Chat/Slack/MessageIdentifierHelperTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Tests\Integration\Infrastructure\Chat\Slack;
 
 use PHPUnit\Framework\TestCase;
-use Slub\Infrastructure\Chat\Slack\MessageIdentifierHelper;
+use Slub\Infrastructure\Chat\Slack\Common\MessageIdentifierHelper;
 
 /**
  * @author    Samir Boulil <samir.boulil@gmail.com>
@@ -17,7 +17,7 @@ class MessageIdentifierHelperTest extends TestCase
      */
     public function it_creates_a_string_out_of_a_workspace_and_channel_and_ts(): void
     {
-        $this->assertEquals('akeneo@general@12345', MessageIdentifierHelper::from('akeneo', 'general', '12345'));
+        $this->assertEquals('akeneo@general@12345', \Slub\Infrastructure\Chat\Slack\Common\MessageIdentifierHelper::from('akeneo', 'general', '12345'));
     }
 
     /**
